Add ERC20 Premints (#48)
Needed to use macros for a lot of things. The main problem is that we need to use separate namespaces for V2 and ERC20 ABI types, because we need to tweak the abi.json and change the names of some types. That's required for signatures to work correctly, but they can't both have the same names.

Experimented with a generic based solution, but ultimately went for the macros because of that.